Adjust settings view
