--- src/components/Settings.tsx
@@ -1,909 +1,0 @@
-import {
-  ChangeEvent,
-  ChangeEventHandler,
-  SyntheticEvent,
-  useEffect,
-  useState,
-} from 'react'
-import Chip, { ChipProps } from '@mui/material/Chip'
-
-import Autocomplete from '@mui/material/Autocomplete'
-import Avatar from '@mui/material/Avatar'
-import Button from '@mui/material/Button'
-import Checkbox from '@mui/material/Checkbox'
-import CircularProgress from '@mui/material/CircularProgress'
-import CloseIcon from '@mui/icons-material/Close'
-import Dialog from '@mui/material/Dialog'
-import DialogActions from '@mui/material/DialogActions'
-import DialogContent from '@mui/material/DialogContent'
-import DialogTitle from '@mui/material/DialogTitle'
-import FirstPageIcon from '@mui/icons-material/FirstPage'
-import FormControlLabel from '@mui/material/FormControlLabel'
-import FormGroup from '@mui/material/FormGroup'
-import Grid from '@mui/material/Grid'
-import GroupIcon from '@mui/icons-material/Group'
-import IconButton from '@mui/material/IconButton'
-import InputAdornment from '@mui/material/InputAdornment'
-import KeyIcon from '@mui/icons-material/Key'
-import VisibilityIcon from '@mui/icons-material/Visibility'
-import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
-import LastPageIcon from '@mui/icons-material/LastPage'
-import Link from '@mui/material/Link'
-import ListIcon from '@mui/icons-material/Ballot'
-import NextIcon from '@mui/icons-material/ChevronRight'
-import OrganizationIcon from '@mui/icons-material/CorporateFare'
-import PipelineIcon from '@mui/icons-material/AccountTree'
-import PreviousIcon from '@mui/icons-material/ChevronLeft'
-import RestoreIcon from '@mui/icons-material/SettingsBackupRestore'
-import SaveIcon from '@mui/icons-material/Save'
-import Stack from '@mui/material/Stack'
-import TextField from '@mui/material/TextField'
-import Typography from '@mui/material/Typography'
-import { dataFetcher } from '../helpers/dataFetcher'
-import { equals } from '../helpers/core'
-import { settingsHandler } from '../helpers/settingsHandler'
-
-const PAGE_SIZE = 24
-
-type PageCursor = {
-  startCursor?: string
-  endCursor?: string
-  page: PageNavigation
-  total?: number
-}
-
-const initialPageCursor: PageCursor = {
-  page: 'NEXT_PAGE',
-}
-
-function usePagination() {
-  const [pageCursor, setPageCursor] = useState<PageCursor>(initialPageCursor)
-
-  const reset = () => setPageCursor(initialPageCursor)
-
-  const navigate = (
-    page: PageNavigation,
-    pageable?: Pick<Pageable, 'startCursor' | 'endCursor' | 'total'>,
-  ) =>
-    setPageCursor({
-      startCursor: pageable?.startCursor,
-      endCursor: pageable?.endCursor,
-      page,
-      total: pageable?.total,
-    })
-
-  return { pageCursor, reset, navigate }
-}
-
-type PaginationControlsProps = {
-  hasPreviousPage: boolean
-  hasNextPage: boolean
-  onFirst: VoidFunction
-  onPrev: VoidFunction
-  onNext: VoidFunction
-  onLast: VoidFunction
-}
-function PaginationControls({
-  hasPreviousPage,
-  hasNextPage,
-  onFirst,
-  onPrev,
-  onNext,
-  onLast,
-}: PaginationControlsProps) {
-  return (
-    <Grid container item xs={12} justifyContent="space-around">
-      <Grid container item xs={6} columnGap={2} justifyContent="center">
-        <Grid item xs={3}>
-          <Button
-            onClick={onFirst}
-            disabled={!hasPreviousPage}
-            startIcon={<FirstPageIcon />}
-          >
-            First
-          </Button>
-        </Grid>
-        <Grid item xs={3}>
-          <Button
-            onClick={onPrev}
-            disabled={!hasPreviousPage}
-            startIcon={<PreviousIcon />}
-          >
-            Prev
-          </Button>
-        </Grid>
-      </Grid>
-      <Grid container item xs={6} columnGap={2} justifyContent="center">
-        <Grid item xs={3}>
-          <Button
-            onClick={onNext}
-            disabled={!hasNextPage}
-            endIcon={<NextIcon />}
-          >
-            Next
-          </Button>
-        </Grid>
-        <Grid item xs={3}>
-          <Button
-            onClick={onLast}
-            disabled={!hasNextPage}
-            endIcon={<LastPageIcon />}
-          >
-            Last
-          </Button>
-        </Grid>
-      </Grid>
-    </Grid>
-  )
-}
-
-function getPermissionColor(
-  permission: TeamRepository['permission'],
-): ChipProps['color'] {
-  switch (permission) {
-    case 'ADMIN':
-      return 'error'
-    case 'MAINTAIN':
-      return 'warning'
-    case 'WRITE':
-      return 'info'
-    case 'READ':
-      return 'success'
-    case 'TRIAGE':
-      return 'secondary'
-    default:
-      return 'default'
-  }
-}
-
-type SettingsProps = {
-  isOpen: boolean
-  close: VoidFunction
-  selectedTeamName: string
-  handleReload: (teamName: string, isValidToken: boolean) => void
-  isLoading: boolean
-}
-export default function Settings({
-  isOpen,
-  close,
-  selectedTeamName,
-  handleReload,
-  isLoading,
-}: SettingsProps) {
-  const [githubToken, setGithubToken] = useState(
-    settingsHandler.loadGithubToken() || '',
-  )
-  const [orgName, setOrgName] = useState(settingsHandler.loadOrgName())
-  const [teamNames, setTeamNames] = useState(settingsHandler.loadTeamNames())
-  const [selectedRepositories, setSelectedRepositories] = useState<Set<string>>(
-    new Set(),
-  )
-  const [selectedMembers, setSelectedMembers] = useState<Set<string>>(new Set())
-  const [loadPipelineStatus, setLoadPipelineStatus] = useState(
-    settingsHandler.loadPipelineStatus(),
-  )
-  const [saveKey, setSaveKey] = useState(0)
-
-  useEffect(() => {
-    const savedTeam = settingsHandler.loadTeam(selectedTeamName)
-    setSelectedRepositories(new Set(savedTeam?.repositories))
-    setSelectedMembers(new Set(savedTeam?.members))
-  }, [selectedTeamName])
-
-  const handleGithubTokenChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setGithubToken(event.target.value)
-  }
-
-  const handleOrgNameChange = (_e: SyntheticEvent, newValue: unknown) => {
-    setOrgName((newValue as Organization)?.login || null)
-  }
-
-  const handleTeamNamesChange = (_e: SyntheticEvent, newValues: unknown[]) => {
-    setTeamNames((newValues as Team[]).map(team => team.name))
-  }
-
-  const handleSave = () => {
-    settingsHandler.partialSave({
-      githubToken,
-      orgName: githubToken ? orgName : null,
-      teamNames: githubToken && orgName ? teamNames : [],
-      teams: {
-        [selectedTeamName]: {
-          repositories: Array.from(selectedRepositories),
-          members: Array.from(selectedMembers),
-        },
-      },
-      loadPipelineStatus,
-    })
-
-    setSaveKey(k => k + 1)
-
-    if (!githubToken || !orgName) {
-      setOrgName(null)
-      setTeamNames([])
-    }
-
-    handleReload(
-      teamNames.includes(selectedTeamName) ? selectedTeamName : teamNames[0],
-      Boolean(githubToken) && Boolean(orgName) && teamNames.length > 0,
-    )
-  }
-
-  const handleReset = () => {
-    const savedTeam = settingsHandler.loadTeam(selectedTeamName)
-    setSelectedRepositories(new Set(savedTeam?.repositories))
-    setSelectedMembers(new Set(savedTeam?.members))
-    setGithubToken(settingsHandler.loadGithubToken())
-    setOrgName(settingsHandler.loadOrgName())
-    setTeamNames(settingsHandler.loadTeamNames())
-    setLoadPipelineStatus(settingsHandler.loadPipelineStatus())
-  }
-
-  const savedTeam = settingsHandler.loadTeam(selectedTeamName)
-  const isResetSettingsDisabled =
-    equals(savedTeam?.repositories.length, selectedRepositories.size) &&
-    equals(savedTeam?.members?.length ?? 0, selectedMembers.size) &&
-    settingsHandler.loadGithubToken() === githubToken &&
-    settingsHandler.loadOrgName() === orgName &&
-    JSON.stringify(settingsHandler.loadTeamNames()) ===
-      JSON.stringify(teamNames) &&
-    settingsHandler.loadPipelineStatus() === loadPipelineStatus
-
-  return (
-    <Dialog open={isOpen} onClose={close} scroll="paper" maxWidth="lg">
-      <DialogTitle>
-        <Stack
-          direction="row"
-          justifyContent="space-between"
-          alignItems="flex-start"
-        >
-          Settings
-          <IconButton edge="end" onClick={close}>
-            <CloseIcon />
-          </IconButton>
-        </Stack>
-      </DialogTitle>
-      <DialogContent>
-        <FormGroup>
-          <Grid container>
-            <AuthorizationSetting
-              githubToken={githubToken}
-              handleGithubTokenChange={handleGithubTokenChange}
-            />
-            <OrganizationSetting
-              orgName={orgName}
-              teamNames={teamNames}
-              handleOrgNameChange={handleOrgNameChange}
-              handleTeamNamesChange={handleTeamNamesChange}
-              githubToken={githubToken}
-            />
-            <PipelineStatusSetting
-              loadPipelineStatus={loadPipelineStatus}
-              setLoadPipelineStatus={setLoadPipelineStatus}
-            />
-            <TeamMembersSetting
-              teamName={selectedTeamName}
-              selectedMembers={selectedMembers}
-              setSelectedMembers={setSelectedMembers}
-              saveKey={saveKey}
-            />
-            <TeamRepositoriesSetting
-              teamName={selectedTeamName}
-              selectedRepositories={selectedRepositories}
-              setSelectedRepositories={setSelectedRepositories}
-              saveKey={saveKey}
-            />
-          </Grid>
-        </FormGroup>
-      </DialogContent>
-      <DialogActions>
-        <Button
-          size="large"
-          onClick={handleReset}
-          startIcon={<RestoreIcon />}
-          disabled={isResetSettingsDisabled}
-        >
-          Reset changes
-        </Button>
-        <Button
-          onClick={handleSave}
-          size="large"
-          startIcon={isLoading ? <CircularProgress size={20} /> : <SaveIcon />}
-        >
-          Save changes
-        </Button>
-      </DialogActions>
-    </Dialog>
-  )
-}
-
-type AuthorizationSettingProps = {
-  githubToken: string
-  handleGithubTokenChange: ChangeEventHandler<HTMLInputElement>
-}
-function AuthorizationSetting({
-  githubToken,
-  handleGithubTokenChange,
-}: AuthorizationSettingProps) {
-  const [showToken, setShowToken] = useState(false)
-
-  return (
-    <Grid
-      container
-      item
-      xs={12}
-      justifyContent="space-between"
-      rowGap={2}
-      sx={{ marginBottom: 2 }}
-    >
-      <Grid item xs={12}>
-        <Typography component="div">
-          <Stack direction="row" alignItems="center" sx={{ fontWeight: 800 }}>
-            <KeyIcon sx={{ marginRight: 1 }} />
-            Authorization (
-            <Link
-              href="https://github.com/settings/tokens"
-              target="_blank"
-              rel="noopener"
-            >
-              link
-            </Link>
-            )
-          </Stack>
-        </Typography>
-      </Grid>
-      <Grid item xs={12}>
-        <TextField
-          id="github-token"
-          label="Github token"
-          type={showToken ? 'text' : 'password'}
-          error={!githubToken}
-          autoComplete="off"
-          helperText={'Must include scopes "repo" and "read:org"'}
-          fullWidth
-          value={githubToken}
-          onChange={handleGithubTokenChange}
-          InputProps={{
-            endAdornment: githubToken ? (
-              <InputAdornment position="end">
-                <IconButton
-                  onClick={() => setShowToken(v => !v)}
-                  edge="end"
-                  size="small"
-                >
-                  {showToken ? <VisibilityOffIcon /> : <VisibilityIcon />}
-                </IconButton>
-              </InputAdornment>
-            ) : undefined,
-          }}
-        />
-      </Grid>
-    </Grid>
-  )
-}
-
-type OrganizationSettingProps = {
-  orgName: string | null
-  teamNames: string[]
-  handleOrgNameChange: (_e: SyntheticEvent, newValue: unknown) => void
-  handleTeamNamesChange: (_e: SyntheticEvent, newValues: unknown[]) => void
-  githubToken?: string
-}
-function OrganizationSetting({
-  orgName,
-  teamNames,
-  handleOrgNameChange,
-  handleTeamNamesChange,
-  githubToken,
-}: OrganizationSettingProps) {
-  const [orgOptions, setOrgOptions] = useState<Organization[]>([])
-  const [isOrgOpen, setIsOrgOpen] = useState(false)
-  const [isOrgLoading, setIsOrgLoading] = useState(false)
-  const [orgInputValue, setOrgInputValue] = useState(
-    orgOptions.find(org => org.login === orgName) ||
-      (orgName && { login: orgName, avatarUrl: '', name: orgName }) ||
-      null,
-  )
-
-  const [teamOptions, setTeamOptions] = useState<Team[]>([])
-  const [isTeamOpen, setIsTeamOpen] = useState(false)
-  const [isTeamLoading, setIsTeamLoading] = useState(false)
-
-  const teamsIntersection = teamOptions.filter(option =>
-    teamNames.find(teamName => teamName === option.name),
-  )
-  const initialTeamsInputValue: Team[] = teamNames.map(
-    teamName =>
-      teamsIntersection.find(option => option.name === teamName) || {
-        name: teamName,
-        avatarUrl: '',
-        description: null,
-      },
-  )
-
-  const [teamsInputValue, setTeamsInputValue] = useState(initialTeamsInputValue)
-
-  useEffect(() => {
-    const getOrganizations = async () => {
-      setIsOrgLoading(true)
-      setOrgOptions(await dataFetcher.fetchOrganizations(githubToken))
-      setIsOrgLoading(false)
-    }
-
-    const getTeams = async () => {
-      if (orgName) {
-        setIsTeamLoading(true)
-        setTeamOptions(await dataFetcher.fetchTeams(orgName, githubToken))
-        setIsTeamLoading(false)
-      }
-    }
-
-    setOrgInputValue(
-      orgOptions.find(org => org.login === orgName) ||
-        (orgName && { login: orgName, avatarUrl: '', name: orgName }) ||
-        null,
-    )
-
-    const tempTeamsIntersection = teamOptions.filter(option =>
-      teamNames.find(teamName => teamName === option.name),
-    )
-    const tempTeamsInputValue: Team[] = teamNames.map(
-      teamName =>
-        tempTeamsIntersection.find(option => option.name === teamName) || {
-          name: teamName,
-          avatarUrl: '',
-          description: null,
-        },
-    )
-    setTeamsInputValue(tempTeamsInputValue)
-
-    if (githubToken) {
-      isOrgOpen && orgOptions.length === 0 && getOrganizations()
-
-      isTeamOpen && orgName && teamOptions.length === 0 && getTeams()
-    } else {
-      // remove all loaded options if there is no token
-      orgOptions.length > 0 && setOrgOptions([])
-
-      teamOptions.length > 0 && setTeamOptions([])
-    }
-
-    if (!orgName && teamOptions.length > 0) {
-      setTeamOptions([])
-    }
-  }, [
-    orgName,
-    orgOptions,
-    isOrgOpen,
-    githubToken,
-    isTeamOpen,
-    teamOptions,
-    teamNames,
-  ])
-
-  return (
-    <Grid
-      container
-      item
-      xs={12}
-      justifyContent="space-between"
-      rowGap={2}
-      sx={{ marginBottom: 2 }}
-    >
-      <Grid item xs={12}>
-        <Typography component="div">
-          <Stack direction="row" alignItems="center" sx={{ fontWeight: 800 }}>
-            <OrganizationIcon sx={{ marginRight: 1 }} />
-            Organization
-          </Stack>
-        </Typography>
-      </Grid>
-      <Grid item xs={12}>
-        <Autocomplete
-          id="organization-name"
-          fullWidth
-          value={orgInputValue}
-          onChange={handleOrgNameChange}
-          options={orgOptions}
-          open={isOrgOpen}
-          onOpen={() => setIsOrgOpen(true)}
-          onClose={() => setIsOrgOpen(false)}
-          loading={isOrgLoading}
-          getOptionLabel={option => option.login}
-          sx={{ marginBottom: 1 }}
-          isOptionEqualToValue={(option, value) => option.login === value.login}
-          renderOption={(props, option) => (
-            <li {...props}>
-              <Stack direction="row" alignItems="center" gap={1}>
-                <Avatar src={option.avatarUrl} />
-                <Typography>{option.name}</Typography>
-                <Typography color="text.secondary">({option.login})</Typography>
-              </Stack>
-            </li>
-          )}
-          renderInput={params => (
-            <TextField
-              {...params}
-              id="orgName"
-              label="Select organization"
-              error={!orgName}
-            />
-          )}
-        />
-        <Autocomplete
-          id="team-names"
-          fullWidth
-          multiple
-          disableCloseOnSelect
-          value={teamsInputValue}
-          onChange={handleTeamNamesChange}
-          options={teamOptions}
-          open={isTeamOpen}
-          onOpen={() => setIsTeamOpen(true)}
-          onClose={() => setIsTeamOpen(false)}
-          loading={isTeamLoading}
-          getOptionLabel={option => option.name}
-          isOptionEqualToValue={(option, value) => option.name === value.name}
-          renderOption={(props, option, { selected }) => (
-            <li {...props}>
-              <Checkbox checked={selected} sx={{ marginRight: 1 }} />
-              <Stack
-                direction="row"
-                alignItems="center"
-                gap={1}
-                overflow="hidden"
-              >
-                <Avatar src={option.avatarUrl} />
-                <Typography>{option.name}</Typography>
-                {option.description && (
-                  <Typography color="text.secondary">
-                    {option.description}
-                  </Typography>
-                )}
-              </Stack>
-            </li>
-          )}
-          renderInput={params => (
-            <TextField
-              {...params}
-              id="teamName"
-              label="Select teams"
-              error={teamNames.length === 0}
-            />
-          )}
-        />
-      </Grid>
-    </Grid>
-  )
-}
-
-type PipelineStatusSettingProps = {
-  loadPipelineStatus: boolean
-  setLoadPipelineStatus: (value: boolean) => void
-}
-function PipelineStatusSetting({
-  loadPipelineStatus,
-  setLoadPipelineStatus,
-}: PipelineStatusSettingProps) {
-  return (
-    <Grid
-      container
-      item
-      xs={12}
-      justifyContent="space-between"
-      sx={{ marginBottom: 2 }}
-    >
-      <Grid item xs={12}>
-        <FormControlLabel
-          label={
-            <Stack direction="row" alignItems="center">
-              <PipelineIcon sx={{ marginRight: 1 }} />
-              <Typography sx={{ fontWeight: 800 }}>
-                Load pipeline status for PRs
-              </Typography>
-            </Stack>
-          }
-          control={
-            <Checkbox
-              checked={loadPipelineStatus}
-              onChange={(e: ChangeEvent<HTMLInputElement>) =>
-                setLoadPipelineStatus(e.target.checked)
-              }
-            />
-          }
-        />
-      </Grid>
-    </Grid>
-  )
-}
-
-type TeamRepositoriesSettingProps = {
-  teamName: string
-  selectedRepositories: Set<string>
-  setSelectedRepositories: (newValue: Set<string>) => void
-  saveKey: number
-}
-function TeamRepositoriesSetting({
-  teamName,
-  selectedRepositories,
-  setSelectedRepositories,
-  saveKey,
-}: TeamRepositoriesSettingProps) {
-  const [isLoading, setIsLoading] = useState(false)
-  const [pageable, setPageable] = useState<TeamRepositoryPageable>()
-  const { pageCursor, reset, navigate } = usePagination()
-  const [currentTeam, setCurrentTeam] = useState('')
-
-  useEffect(() => {
-    if (currentTeam !== teamName) {
-      setCurrentTeam(teamName)
-      reset()
-      return
-    }
-
-    const canFetch =
-      Boolean(settingsHandler.loadGithubToken()) &&
-      Boolean(settingsHandler.loadOrgName()) &&
-      settingsHandler.loadTeamNames().length > 0
-
-    if (!canFetch) {
-      setPageable(undefined)
-      return
-    }
-
-    const fetch = async () => {
-      setIsLoading(true)
-      setPageable(
-        await dataFetcher
-          .fetchTeamRepositories(
-            settingsHandler.loadOrgName() || '',
-            teamName,
-            pageCursor.page,
-            PAGE_SIZE,
-            pageCursor.startCursor,
-            pageCursor.endCursor,
-            pageCursor.total,
-          )
-          .catch(error => {
-            setIsLoading(false)
-            throw error
-          }),
-      )
-      setIsLoading(false)
-    }
-
-    fetch()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [teamName, pageCursor, currentTeam, saveKey])
-
-  const handleUnselectAllRepositories = () => {
-    setSelectedRepositories(new Set())
-  }
-
-  const handleSelectAllRepositoriesOnPage = () => {
-    pageable?.teamRepositories.forEach(repo =>
-      selectedRepositories.add(repo.nameWithOwner),
-    )
-    setSelectedRepositories(new Set(selectedRepositories))
-  }
-
-  const handleRepositoryClick = (nameWithOwner: string) => {
-    if (selectedRepositories.has(nameWithOwner)) {
-      selectedRepositories.delete(nameWithOwner)
-    } else {
-      selectedRepositories.add(nameWithOwner)
-    }
-    setSelectedRepositories(new Set(selectedRepositories))
-  }
-
-  const selectedSize = selectedRepositories.size
-
-  return (
-    <>
-      <Grid container item xs={12} justifyContent="space-between">
-        <Grid item>
-          <Typography
-            component="div"
-            color={!teamName ? 'text.secondary' : undefined}
-          >
-            <Stack direction="row" alignItems="center" sx={{ fontWeight: 800 }}>
-              <ListIcon sx={{ marginRight: 1 }} />
-              {teamName
-                ? `Repositories of ${teamName} (${
-                    pageable?.total || 0
-                  } total${`, ${
-                    selectedSize > 0 && selectedSize < (pageable?.total || 0)
-                      ? selectedSize
-                      : 'none'
-                  } selected`})`
-                : 'Repositories'}
-              {isLoading && (
-                <CircularProgress size={20} sx={{ marginLeft: 1 }} />
-              )}
-            </Stack>
-          </Typography>
-        </Grid>
-        {pageable && (
-          <Grid item>
-            <Button size="small" onClick={handleSelectAllRepositoriesOnPage}>
-              Select all
-            </Button>
-            <Button
-              size="small"
-              onClick={handleUnselectAllRepositories}
-              disabled={!selectedRepositories.size}
-            >
-              Unselect all
-            </Button>
-          </Grid>
-        )}
-      </Grid>
-      {pageable?.teamRepositories.map(({ name, nameWithOwner, permission }) => (
-        <Grid item key={nameWithOwner} lg={4} sm={6} xs={12}>
-          <FormControlLabel
-            label={
-              <>
-                <Stack direction="row" alignItems="center">
-                  {name}
-                  <Chip
-                    color={getPermissionColor(permission)}
-                    label={permission}
-                    size="small"
-                    variant="outlined"
-                    sx={{ marginLeft: 1 }}
-                  />
-                </Stack>
-              </>
-            }
-            control={
-              <Checkbox
-                checked={selectedRepositories.has(nameWithOwner)}
-                onClick={() => handleRepositoryClick(nameWithOwner)}
-              />
-            }
-          />
-        </Grid>
-      )) || (
-        <Grid item marginTop={2}>
-          <Typography color="text.secondary">No data available</Typography>
-        </Grid>
-      )}
-      {pageable && pageable.total > PAGE_SIZE && (
-        <PaginationControls
-          hasPreviousPage={pageable.hasPreviousPage}
-          hasNextPage={pageable.hasNextPage}
-          onFirst={() => navigate('FIRST_PAGE', pageable)}
-          onPrev={() => navigate('PREVIOUS_PAGE', pageable)}
-          onNext={() => navigate('NEXT_PAGE', pageable)}
-          onLast={() => navigate('LAST_PAGE', pageable)}
-        />
-      )}
-    </>
-  )
-}
-
-type TeamMembersSettingProps = {
-  teamName: string
-  selectedMembers: Set<string>
-  setSelectedMembers: (newValue: Set<string>) => void
-  saveKey: number
-}
-function TeamMembersSetting({
-  teamName,
-  selectedMembers,
-  setSelectedMembers,
-  saveKey,
-}: TeamMembersSettingProps) {
-  const [isLoading, setIsLoading] = useState(false)
-  const [members, setMembers] = useState<User[]>([])
-
-  useEffect(() => {
-    const canFetch =
-      Boolean(settingsHandler.loadGithubToken()) &&
-      Boolean(settingsHandler.loadOrgName()) &&
-      settingsHandler.loadTeamNames().length > 0 &&
-      Boolean(teamName)
-
-    if (!canFetch) {
-      setMembers([])
-      return
-    }
-
-    const fetch = async () => {
-      setIsLoading(true)
-      try {
-        const result = await dataFetcher.fetchTeamUsers(
-          settingsHandler.loadOrgName() || '',
-          teamName,
-        )
-        setMembers(result)
-      } finally {
-        setIsLoading(false)
-      }
-    }
-
-    fetch()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [teamName, saveKey])
-
-  const handleToggleMember = (login: string) => {
-    const next = new Set(selectedMembers)
-    if (next.has(login)) {
-      next.delete(login)
-    } else {
-      next.add(login)
-    }
-    setSelectedMembers(next)
-  }
-
-  const handleUnselectAll = () => setSelectedMembers(new Set())
-  const handleSelectAll = () =>
-    setSelectedMembers(new Set(members.map(m => m.login)))
-
-  const selectedSize = selectedMembers.size
-
-  return (
-    <>
-      <Grid container item xs={12} justifyContent="space-between">
-        <Grid item>
-          <Typography
-            component="div"
-            color={!teamName ? 'text.secondary' : undefined}
-          >
-            <Stack direction="row" alignItems="center" sx={{ fontWeight: 800 }}>
-              <GroupIcon sx={{ marginRight: 1 }} />
-              {teamName
-                ? `Members of ${teamName} (${members.length} total${
-                    selectedSize > 0 && selectedSize < members.length
-                      ? `, ${selectedSize} selected`
-                      : ', none selected'
-                  })`
-                : 'Members'}
-              {isLoading && (
-                <CircularProgress size={20} sx={{ marginLeft: 1 }} />
-              )}
-            </Stack>
-          </Typography>
-        </Grid>
-        {members.length > 0 && (
-          <Grid item>
-            <Button size="small" onClick={handleSelectAll}>
-              Select all
-            </Button>
-            <Button
-              size="small"
-              onClick={handleUnselectAll}
-              disabled={!selectedMembers.size}
-            >
-              Unselect all
-            </Button>
-          </Grid>
-        )}
-      </Grid>
-      {members.length > 0 ? (
-        members.map(({ login, avatarUrl }) => (
-          <Grid item key={login} lg={4} sm={6} xs={12}>
-            <FormControlLabel
-              label={
-                <Stack direction="row" alignItems="center" gap={1}>
-                  <Avatar src={avatarUrl} sx={{ width: 24, height: 24 }} />
-                  <Typography>{login}</Typography>
-                </Stack>
-              }
-              control={
-                <Checkbox
-                  checked={selectedMembers.has(login)}
-                  onClick={() => handleToggleMember(login)}
-                />
-              }
-            />
-          </Grid>
-        ))
-      ) : (
-        <Grid item marginTop={2}>
-          <Typography color="text.secondary">No data available</Typography>
-        </Grid>
-      )}
-    </>
-  )
-}
