+++ src/helpers/queryCache.ts
@@ -1,0 +1,13 @@
+const store = new Map<string, unknown>()
+
+export const queryCache = {
+  get<T>(key: string): T | undefined {
+    return store.get(key) as T | undefined
+  },
+  set<T>(key: string, value: T): void {
+    store.set(key, value)
+  },
+  clear(): void {
+    store.clear()
+  },
+}
